requirements.txt
Django>=4.0,<5.0
djangorestframework
celery
redis
psycopg2
djangorestframework-simplejwt
gunicorn
python-dotenv
